[addons][audioencoder] rename audio_encoder.h to audioencoder.h
With the newly planned sandbox / API level system it becomes necessary to have the "C" files matching its namespace.

As an example, a "C" function in audioencoder must use "kodi_addons_audioencoder_some_function()" as the name, with which the "_" acts as a namespace divider to begin with, if now, as before, it would see "audio" and "encoder" as separate areas.

Advantageous for the automatic creation script, otherwise problems can arise for him.
