+++ xbmc/addons/kodi-dev-kit/include/kodi/c-api/addon-instance/CMakeLists.txt
@@ -3,7 +3,7 @@
 
 set(HEADERS
   audiodecoder.h
-  audio_encoder.h
+  audioencoder.h
   game.h
   image_decoder.h
   inputstream.h
