+++ xbmc/addons/kodi-dev-kit/include/kodi/versions.h
@@ -94,7 +94,7 @@
 #define ADDON_INSTANCE_VERSION_AUDIOENCODER           "2.1.0"
 #define ADDON_INSTANCE_VERSION_AUDIOENCODER_MIN       "2.1.0"
 #define ADDON_INSTANCE_VERSION_AUDIOENCODER_XML_ID    "kodi.binary.instance.audioencoder"
-#define ADDON_INSTANCE_VERSION_AUDIOENCODER_DEPENDS   "c-api/addon-instance/audio_encoder.h" \
+#define ADDON_INSTANCE_VERSION_AUDIOENCODER_DEPENDS   "c-api/addon-instance/audioencoder.h" \
                                                       "addon-instance/AudioEncoder.h"
 
 #define ADDON_INSTANCE_VERSION_GAME                   "2.2.0"
